文生图组件仅更改到V1版本
