+++ python/core/components/v2/__init__.py
@@ -22,7 +22,6 @@ from .qrcode_ocr.component import QRcodeOCR
 from .handwrite_ocr.component import HandwriteOCR
 from .mix_card_ocr.component import MixCardOCR
 from .table_ocr.component import TableOCR
-from .text_to_image.component import Text2Image
 from .llms.style_writing.component import StyleWriting
 
 __V2_COMPONENTS__ = [
@@ -36,6 +35,5 @@ __V2_COMPONENTS__ = [
     "HandwriteOCR",
     "MixCardOCR",
     "TableOCR",
-    "Text2Image",
     "StyleWriting"
 ] # NOQA
